requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
apturl==0.5.2
async-timeout==4.0.3
attrs==21.2.0
azure-core==1.29.6
azure-functions==1.17.0
azure-functions-durable==1.2.8
azure-identity==1.15.0
azure-storage-blob==12.19.0
bcrypt==4.0.1
blinker==1.7.0
Brlapi==0.8.3
Brotli==1.1.0
capstone==5.0.1
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorama==0.4.4
colored-traceback==0.3.0
command-not-found==0.3
ConfigArgParse==1.7
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.2.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
frozenlist==1.4.1
furl==2.1.3
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.3
gunicorn==22.0.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
intervaltree==3.1.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
locust==2.20.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.0
Markdown==3.3.6
MarkupSafe==2.1.3
more-itertools==8.10.0
msal==1.26.0
msal-extensions==1.1.0
msgpack==1.0.7
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
orderedmultidict==1.0.1
packaging==23.2
paramiko==3.3.1
pexpect==4.8.0
Pillow==9.0.1
plumbum==1.8.2
portalocker==2.8.2
protobuf==3.12.4
psutil==5.9.7
ptyprocess==0.7.0
pwntools==4.11.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pyelftools==0.30
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
pyperclip3==0.4.1
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
rdflib==7.0.0
reportlab==3.6.8
requests==2.25.1
ROPGadget==7.4
roundrobin==0.0.4
rpyc==5.3.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
systemd-python==234
texttable==1.6.4
ttystatus==0.38
typing_extensions==4.9.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unicorn==2.0.1.post1
urllib3==1.26.5
urwid==2.1.2
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
zope.event==5.0
zope.interface==6.1